优化初音技能说明和实际效果 #216
