+++ resource/addon_english.txt
@@ -141,30 +141,39 @@
 
 		// 初音未来
 		"npc_dota_hero_meepo"					"Miku"
+		"npc_dota_hero_meepo_npedesc1"		"Hatsune Miku Princess"
 		"DOTA_Tooltip_ability_get_down" "GET DOWN!!!"
-        "DOTA_Tooltip_ability_get_down_Description" "Uses his own secret dance, inflicting a blow from the hand to each enemy around him every half second for the entire duration of the dance. The blows do not pass through the invulnerability to magic"
-        "DOTA_Tooltip_ability_get_down_sand_storm_radius" "Radius:"
-        "DOTA_Tooltip_ability_get_down_sand_storm_damage" "BonusMagicDamagePerHit:"
+        "DOTA_Tooltip_ability_get_down_Description" "Uses his own secret dance, inflicting a Normal Attack to each enemy around him every second for the entire duration of the dance. Deals bonus magic damage per hit. The blows do not pass through the invulnerability to magic."
+        "DOTA_Tooltip_ability_get_down_dance_radius" "Radius:"
+        "DOTA_Tooltip_ability_get_down_dance_damage" "Bonus Magic Damage Per Hit:"
+		"DOTA_Tooltip_ability_get_down_Lore"  "Is this enough?"
+		"DOTA_Tooltip_ability_get_down_scepter_description"    "Get new ability <font color='#FFFFFF'>GET DOWN</font>.<br/>Hatsune Miku's Secret Dance, inflicting a Normal Attack to enemy around and bonus magic damage per second. Ability has same level with <font color='#FFFFFF'>Chibi Miku</font>"
         "DOTA_Tooltip_ability_chibi_hit" "CHIBI STRIKE"
-        "DOTA_Tooltip_ability_chibi_hit_Description" "Deals a devastating blow to the target enemy, dealing a colossal amount of damage to them. If they survive, they are stunned for 2.5 seconds"
+        "DOTA_Tooltip_ability_chibi_hit_Description" "Deals a devastating blow to the target enemy, dealing a colossal amount of damage and stunning them."
         "DOTA_Tooltip_ability_chibi_hit_AbilityDamage" "Damage:"
+        "DOTA_Tooltip_ability_chibi_hit_stun_duration" "Stun duration:"
         "DOTA_Tooltip_ability_chibi_monster" "Chibi Miku"
-        "DOTA_Tooltip_ability_chibi_monster_Description" "Transforms into a large Chibi version of herself, gaining additional stats and a new ability.<br/>Increases the damage of Popipo's abilities."
+        "DOTA_Tooltip_ability_chibi_monster_Description" "Transforms into a large Chibi version of herself, gaining additional stats and a new ability.<br/>Increases the damage of <font color='#FFFFFF'>Popipo</font>.<br/>Makes <font color='#FFFFFF'>Liu-kick</font> can stun enemy."
         "DOTA_Tooltip_ability_chibi_monster_damage" "BonusDamage:"
         "DOTA_Tooltip_ability_chibi_monster_agility" "BonusAgi:"
         "DOTA_Tooltip_ability_chibi_monster_hp" "BonusHp:"
         "DOTA_Tooltip_ability_chibi_monster_fixed_duration" "Duration:"
         "DOTA_Tooltip_ability_miku_dance" "Dancing"
-        "DOTA_Tooltip_ability_miku_dance_Description" "Begins to sing and dance, dealing damage to all enemies around, and stunning them for 1 second at intervals of 1.5"
-        "DOTA_Tooltip_ability_miku_dance_sand_storm_radius" "Radius:"
-        "DOTA_Tooltip_ability_miku_dance_sand_storm_damage" "DamagePerTick:"
+        "DOTA_Tooltip_ability_miku_dance_Description" "Begins to sing and dance, inflicting a Normal Attack to enemy around and dealing bonus magic damage to all enemies around, and stunning them for %stun_duration% second at intervals of %attack_interval%.<br/>Hatsune Miku takes less damage while dancing."
+        "DOTA_Tooltip_ability_miku_dance_dance_radius" "Radius:"
+        "DOTA_Tooltip_ability_miku_dance_dance_damage" "DamagePerTick:"
+        "DOTA_Tooltip_ability_miku_dance_damage_reduction" "%Damage reduction:"
         "DOTA_Tooltip_ability_liu_kick" "Liu-kick"
-        "DOTA_Tooltip_ability_liu_kick_Description" "Performs a long leap from his feet to the target, inflicting some amount of magic damage and a blow from the hand"
+        "DOTA_Tooltip_ability_liu_kick_Description" "Performs a long leap from his feet to the target, inflicting some amount of magic damage, knockback, stun and a Normal Attack.<br/>Stuns the enemy when in Chibi status."
         "DOTA_Tooltip_ability_liu_kick_damage" "Damage:"
+        "DOTA_Tooltip_ability_liu_kick_range" "Caste Range:"
+        "DOTA_Tooltip_ability_liu_kick_stun_duration" "Stun duration in Chibi status:"
+		"DOTA_Tooltip_ability_liu_kick_Lore" "Miku~miku~Kick"
         "DOTA_Tooltip_ability_pantsushot" "PantsuShot"
-        "DOTA_Tooltip_ability_pantsushot_Description" "Has a chance to inflict a crushing kick from a 720 degree roundabout, dealing critical damage, stunning the enemy for 0.3 seconds"
-        "DOTA_Tooltip_ability_pantsushot_crit_chance" "Chance:"
+        "DOTA_Tooltip_ability_pantsushot_Description" "Has a chance to inflict a crushing kick from a 720 degree roundabout, dealing critical damage, stunning the enemy."
+        "DOTA_Tooltip_ability_pantsushot_crit_chance" "%Chance:"
         "DOTA_Tooltip_ability_pantsushot_crit_bonus" "%Crit:"
+        "DOTA_Tooltip_ability_pantsushot_stun_duration" "Stun duration:"
         "DOTA_Tooltip_ability_miku_hadouken" "Popipo"
         "DOTA_Tooltip_ability_miku_hadouken_Description" "Creates a musical wave directed in the specified direction, damaging all enemies in its path"
         "DOTA_Tooltip_ability_miku_hadouken_damage" "Damage:"
