+++ resource/addon_schinese.txt
@@ -143,13 +143,15 @@
 		"npc_dota_hero_meepo"					"初音未来"
 		"npc_dota_hero_meepo_npedesc1"		"世界第一的公主殿下"
 		"DOTA_Tooltip_ability_get_down" "GET DOWN!!!"
-        "DOTA_Tooltip_ability_get_down_Description" "使用初音未来自己的秘密舞蹈，在整个舞蹈过程中，每半秒对周围敌人造成一次伤害不无视魔免"
-        "DOTA_Tooltip_ability_get_down_sand_storm_radius" "范围:"
-        "DOTA_Tooltip_ability_get_down_sand_storm_dmage" "每次命中增加魔法伤害:"
+        "DOTA_Tooltip_ability_get_down_Description" "使用初音未来自己的秘密舞蹈，在整个舞蹈过程中，每%attack_interval%秒对周围敌人进行一次普通攻击，并且附加额外魔法伤害。"
+        "DOTA_Tooltip_ability_get_down_dance_radius" "范围:"
+        "DOTA_Tooltip_ability_get_down_dance_damage" "每次攻击附加魔法伤害:"
 		"DOTA_Tooltip_ability_get_down_Lore"  "这样一来就满足了吗？"
+		"DOTA_Tooltip_ability_get_down_scepter_description"    "获得新技能 <font color='#FFFFFF'>GET DOWN</font>。<br/>初音未来的秘密舞蹈，对周围敌人造成伤害。技能等级随着 <font color='#FFFFFF'>甩葱形态</font> 增加。"
         "DOTA_Tooltip_ability_chibi_hit" "葱击"
-        "DOTA_Tooltip_ability_chibi_hit_Description" "对目标敌人造成毁灭性打击，对他们造成巨大伤害。如果他们活下来，会被击晕2.5秒"
+        "DOTA_Tooltip_ability_chibi_hit_Description" "对目标敌人造成毁灭性打击，对他们造成巨大伤害，并且击晕。"
         "DOTA_Tooltip_ability_chibi_hit_AbilityDamage" "伤害:"
+        "DOTA_Tooltip_ability_chibi_hit_stun_duration" "眩晕时间:"
         "DOTA_Tooltip_ability_chibi_monster" "甩葱形态"
         "DOTA_Tooltip_ability_chibi_monster_Description" "初音变身成甩葱形态，获得额外的属性和技能。<br/>变身时增加技能蔬菜汁Popipo的伤害。"
         "DOTA_Tooltip_ability_chibi_monster_damage" "获得伤害:"
@@ -158,25 +160,28 @@
         "DOTA_Tooltip_ability_chibi_monster_fixed_duration" "持续时间:"
 		"DOTA_Tooltip_ability_chibi_monster_Lore" "你一定听过甩葱歌吧！"
         "DOTA_Tooltip_ability_miku_dance" "跳舞"
-        "DOTA_Tooltip_ability_miku_dance_Description" "开始唱歌跳舞，对周围所有敌人造成伤害，每1.5秒眩晕周围敌人一秒"
-        "DOTA_Tooltip_ability_miku_dance_sand_storm_radius" "范围:"
-        "DOTA_Tooltip_ability_miku_dance_sand_storm_damage" "每次攻击伤害:"
+        "DOTA_Tooltip_ability_miku_dance_Description" "开始唱歌跳舞，每%attack_interval%秒对周围敌人进行一次普通攻击，并眩晕周围敌人%stun_duration%秒，附加额外魔法伤害。<br/>跳舞时初音收到的伤害减少。"
+        "DOTA_Tooltip_ability_miku_dance_dance_radius" "范围:"
+        "DOTA_Tooltip_ability_miku_dance_dance_damage" "每次攻击伤害:"
+        "DOTA_Tooltip_ability_miku_dance_damage_reduction" "%伤害减免:"
 		"DOTA_Tooltip_ability_miku_dance_Lore"  "我是世界第一的公主殿下，唱、跳、Rap都不在话下"
         "DOTA_Tooltip_ability_liu_kick" "初音飞踢"
-        "DOTA_Tooltip_ability_liu_kick_Description" "踢向目标，造成一定的魔法伤害和击退"
+        "DOTA_Tooltip_ability_liu_kick_Description" "踢向目标将其击退，对其与周围敌人造成一定的魔法伤害，并附加一次普通攻击。<br/>变身时对敌人造成眩晕。"
         "DOTA_Tooltip_ability_liu_kick_damage" "伤害:"
+        "DOTA_Tooltip_ability_liu_kick_range" "施法距离:"
+        "DOTA_Tooltip_ability_liu_kick_stun_duration" "变身时眩晕时间:"
 		"DOTA_Tooltip_ability_liu_kick_Lore" "把你mikumiku掉！"
         "DOTA_Tooltip_ability_pantsushot" "脑浆炸裂女孩"
-        "DOTA_Tooltip_ability_pantsushot_Description" "有几率造成致命一击,使敌人晕眩0.3秒"
-        "DOTA_Tooltip_ability_pantsushot_crit_chance" "几率:"
-        "DOTA_Tooltip_ability_pantsushot_crit_bonus" "%暴击."
+        "DOTA_Tooltip_ability_pantsushot_Description" "有几率造成致命一击,使敌人晕眩。"
+        "DOTA_Tooltip_ability_pantsushot_crit_chance" "%几率:"
+        "DOTA_Tooltip_ability_pantsushot_crit_bonus" "%暴击:"
+        "DOTA_Tooltip_ability_pantsushot_stun_duration" "眩晕时间:"
 		"DOTA_Tooltip_ability_pantsushot_Lore" "Nyanyanyanyanyanyanya!"
 		"DOTA_Tooltip_ability_miku_hadouken"           "蔬菜汁Popipo"
 		"DOTA_Tooltip_ability_miku_hadouken_Description" "产生一个指向指定方向的Popipo音乐波，伤害其路径上的所有敌人"
         "DOTA_Tooltip_ability_miku_hadouken_damage"         "伤害:"
         "DOTA_Tooltip_ability_miku_hadouken_damage_calne"   "变身时伤害:"
 		"DOTA_Tooltip_ability_miku_hadouken_Lore"           "香醇浓厚的蔬菜果汁！香软绵柔的蔬菜果汁！"
-		"DOTA_Tooltip_ability_get_down_scepter_description"    "初音未来的秘密舞蹈，很鬼畜"
 
 		// 双头龙
         "DOTA_Tooltip_ability_jakiro_macropyre_scepter_description"			"烈焰焚身的距离+200，持续时间+5秒，并且变为纯粹伤害，无视技能免疫。"
