+++ scripts/vscripts/heroes/hero_miku/pantsushot.lua
@@ -1,5 +1,5 @@
 pantsushot = class({})
-LinkLuaModifier( "modifier_pantsushot", "modifiers/modifier_pantsushot", LUA_MODIFIER_MOTION_NONE )
+LinkLuaModifier( "modifier_pantsushot", "modifiers/hero_miku/modifier_pantsushot", LUA_MODIFIER_MOTION_NONE )
 LinkLuaModifier( "modifier_generic_stunned_lua", "modifiers/modifier_generic_stunned_lua", LUA_MODIFIER_MOTION_NONE )
 --------------------------------------------------------------------------------
 -- Passive Modifier
